Fix: Load Repositories button disabling button
